Dont output types for libraries we aren't tracing

+++ data/typing/numpy.lib.stride_tricks.py
@@ -3,7 +3,7 @@ from typing import *
 
 def as_strided(
     x: numpy.ndarray,
-    shape: Tuple[(Union[(int, numpy.int64)], ...)],
-    strides: Tuple[(Union[(int, numpy.int64)], ...)] = ...,
+    shape: Tuple[(Union[(numpy.int64, int)], ...)],
+    strides: Tuple[(Union[(numpy.int64, int)], ...)] = ...,
 ):
     "usage.skimage: 11, usage.xarray: 8, usage.sklearn: 6"
